auth/requirements.txt
Django
djangorestframework
djoser
djangorestframework_simplejwt
mysqlclient
